Deconstructs css a bit so that I can expand into a Dashboard Component to contain my various views
