+++ src/components/DashboardFooter.jsx
@@ -1,0 +1,14 @@
+import React, { Component } from "react";
+import '../resources/css/footer.css';
+
+
+export default class Footer extends Component {
+
+    render() {
+        return (
+            <div className='dashboardFooter'>
+                <h2> Footer </h2>
+            </div>
+        );
+    }
+}
