+++ src/components/DashboardHeader.jsx
@@ -1,0 +1,14 @@
+import React, { Component } from "react";
+import '../resources/css/footer.css';
+
+
+export default class Footer extends Component {
+
+    render() {
+        return (
+            <div className='dashboardHeader'>
+                <h2> Character </h2>
+            </div>
+        );
+    }
+}
