+++ src/resources/css/character.css
@@ -1,6 +1,6 @@
 .characterBuilder {
   margin: auto;
-  min-height: 445px;
+  min-height: inherit;
   /* border: 2px solid var(--component-bg-color); */
   background-color: var(--main-alt-bg-color);
   display: flex;
@@ -19,7 +19,6 @@
   background-color: var(--component-bg-color);
   text-align: center;
   padding: 8px;
-  border-bottom: 2px solid var(--component-border-color);
 }
 .characterBuilder h3{
   color: var(--component-text-color);
