--- src/utils/functionalComponent/Dashboard.js
@@ -1,16 +1,0 @@
-import React from "react";
-import CharacterView from './CharacterView';
-import '../../resources/css/dashboard.css'
-
-function Dashboard() {
-
-    return (
-        <div className="dashboard">
-            <CharacterView />
-        </div>
-    );
-}
-
-export default Dashboard;
-
-
