+++ src/utils/functionalComponent/DashboardNavigation.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+export default function DashboardNavigation() {
+
+    return (
+        <div className='dashboardNavigation'>
+            <div>
+                HERE
+            </div>
+        </div>
+    );
+}
+
+
+
+
