+++ src/utils/functionalComponent/DashboardView.js
@@ -1,0 +1,14 @@
+import React from "react";
+import CharacterView from './CharacterView';
+
+import '../../resources/css/dashboard.css';
+
+function Dashboard() {
+    return (
+        <CharacterView />
+    );
+}
+
+export default Dashboard;
+
+
