+++ src/utils/functionalComponent/Main.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Dashboard from "./Dashboard";
+import Dashboard from "../../components/Dashboard";
 import Landing from "../../components/Landing"
 import { useAuth } from '../useAuth';
